requirements.txt
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-search-nspkg==3.0.1
azure-cognitiveservices-search-websearch==1.0.0
azure-common==1.1.20
azure-nspkg==3.0.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msrest==0.6.6
oauthlib==3.0.1
pyodbc==4.0.26
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.2
Werkzeug==0.15.4
